Fix lint and tests
