+++ packages/studio-base/src/components/HelpSidebar/index.stories.tsx
@@ -11,12 +11,18 @@
 //   found at http://www.apache.org/licenses/LICENSE-2.0
 //   You may not use this file except in compliance with the License.
 
+import { useCallback, useRef } from "react";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 
+import { useShallowMemo } from "@foxglove/hooks";
+import HelpInfoContext, {
+  IHelpInfo,
+  HelpInfo,
+} from "@foxglove/studio-base/context/HelpInfoContext";
 import { PanelCatalog, PanelInfo } from "@foxglove/studio-base/context/PanelCatalogContext";
 import MockCurrentLayoutProvider from "@foxglove/studio-base/providers/CurrentLayoutProvider/MockCurrentLayoutProvider";
-import MockHelpInfoProvider from "@foxglove/studio-base/providers/HelpInfoProvider/MockHelpInfoProvider";
+import { DEFAULT_HELP_INFO } from "@foxglove/studio-base/providers/HelpInfoProvider";
 import PanelSetup from "@foxglove/studio-base/stories/PanelSetup";
 import { PanelConfigSchemaEntry } from "@foxglove/studio-base/types/panels";
 
@@ -52,6 +58,37 @@ class MockPanelCatalog implements PanelCatalog {
   getPanelByType(type: string): PanelInfo | undefined {
     return allPanels.find((panel) => panel.preconfigured !== true && panel.type === type);
   }
+}
+
+function MockHelpInfoProvider({
+  isEmpty,
+  children,
+}: {
+  isEmpty?: boolean;
+  children: React.ReactNode;
+}): JSX.Element {
+  const isHelpInfoSet = isEmpty == undefined ? false : isEmpty;
+  const helpInfo = useRef<HelpInfo>(
+    isHelpInfoSet ? DEFAULT_HELP_INFO : { title: "Some title", content: <>Some help content</> },
+  );
+  const helpInfoListeners = useRef(new Set<(_: HelpInfo) => void>());
+
+  const getHelpInfo = useCallback((): HelpInfo => helpInfo.current, []);
+  const setHelpInfo = useCallback((info: HelpInfo): void => {
+    helpInfo.current = info;
+    for (const listener of [...helpInfoListeners.current]) {
+      listener(helpInfo.current);
+    }
+  }, []);
+
+  const value: IHelpInfo = useShallowMemo({
+    getHelpInfo,
+    setHelpInfo,
+    addHelpInfoListener: useCallback(() => {}, []),
+    removeHelpInfoListener: useCallback(() => {}, []),
+  });
+
+  return <HelpInfoContext.Provider value={value}>{children}</HelpInfoContext.Provider>;
 }
 
 export function Home(): JSX.Element {
